+++ admin/meta-box-subscription-info.php
@@ -4,8 +4,6 @@ $post_id = get_the_ID();
 
 $subscription = get_pronamic_subscription( $post_id );
 
-$payment = $subscription->get_first_payment();
-
 ?>
 <table class="form-table">
 	<tr>
@@ -65,50 +63,14 @@ $payment = $subscription->get_first_payment();
 	</tr>
 	<tr>
 		<th scope="row">
-			<?php esc_html_e( 'Start date', 'pronamic_ideal' ); ?>
-		</th>
-		<td>
-			<?php
-
-			$start_date = get_date_from_gmt( $subscription->get_start_date()->format( 'Y-m-d H:i:s' ) );
-
-			echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $start_date ) ) );
-
-			?>
-		</td>
-	</tr>
-	<tr>
-		<th scope="row">
-			<?php esc_html_e( 'Expiration date', 'pronamic_ideal' ); ?>
-		</th>
-		<td>
-			<?php
-
-			$expiration_date = $subscription->get_expiration_date();
-
-			if ( $expiration_date ) {
-				$date = get_date_from_gmt( $expiration_date->format( 'Y-m-d H:i:s' ) );
-
-				echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $date ) ) );
-			} else {
-				echo esc_html( __( 'Not applicable', 'pronamic_ideal' ) );
-			}
-
-			?>
-		</td>
-	</tr>
-	<tr>
-		<th scope="row">
 			<?php esc_html_e( 'First payment', 'pronamic_ideal' ); ?>
 		</th>
 		<td>
 			<?php
 
-			$first_date = $subscription->get_first_payment_date();
+			$first_date = get_date_from_gmt( $subscription->get_first_payment_date()->format( 'Y-m-d H:i:s' ) );
 
-			$date = get_date_from_gmt( $first_date->format( 'Y-m-d H:i:s' ) );
-
-			echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $date ) ) );
+			echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $first_date ) ) );
 
 			?>
 		</td>
@@ -120,12 +82,12 @@ $payment = $subscription->get_first_payment();
 		<td>
 			<?php
 
-			$next_payment = $subscription->get_next_payment_datetime();
+			$next_payment = $subscription->get_next_payment_date();
 
 			if ( $next_payment ) {
-				$local_date = get_date_from_gmt( $next_payment->format( 'Y-m-d H:i:s' ) );
+				$next_date = get_date_from_gmt( $next_payment->format( 'Y-m-d H:i:s' ) );
 
-				echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $local_date ) ) );
+				echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $next_date ) ) );
 			} else {
 				echo esc_html( __( 'None', 'pronamic_ideal' ) );
 			}
@@ -140,14 +102,40 @@ $payment = $subscription->get_first_payment();
 		<td>
 			<?php
 
-			$final_payment = $subscription->get_final_payment_datetime();
+			$final_payment = $subscription->get_final_payment_date();
 
 			if ( $final_payment ) {
-				$local_date = get_date_from_gmt( $final_payment->format( 'Y-m-d H:i:s' ) );
+				$final_date = get_date_from_gmt( $final_payment->format( 'Y-m-d H:i:s' ) );
 
-				echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $local_date ) ) );
-			} else {
-				echo esc_html( __( 'Not applicable', 'pronamic_ideal' ) );
+				echo esc_html( date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $final_date ) ) );
+			}
+
+			?>
+		</td>
+	</tr>
+	<tr>
+		<th scope="row">
+			<?php esc_html_e( 'Current period', 'pronamic_ideal' ); ?>
+		</th>
+		<td>
+			<?php
+
+			$start_date  = $subscription->get_start_date();
+			$expiry_date = $subscription->get_expiry_date();
+
+			if ( $start_date && $expiry_date ) {
+				$start_date  = get_date_from_gmt( $start_date->format( 'Y-m-d H:i:s' ) );
+				$expiry_date = get_date_from_gmt( $expiry_date->format( 'Y-m-d H:i:s' ) );
+
+				echo esc_html( sprintf(
+					'%s – %s',
+					date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $start_date ) ),
+					date_i18n( __( 'l jS \o\f F Y, h:ia', 'pronamic_ideal' ), strtotime( $expiry_date ) )
+				) );
+			}
+
+			if ( Pronamic_WP_Pay_Statuses::COMPLETED === $subscription->get_status() ) {
+				echo esc_html( sprintf( __( 'None (subscription completed)', 'pronamic_ideal' ) ) );
 			}
 
 			?>
@@ -176,8 +164,10 @@ $payment = $subscription->get_first_payment();
 		<td>
 			<?php
 
-			if ( $payment ) {
-				echo $payment->get_source_text(); //xss ok
+			$first_payment = $subscription->get_first_payment();
+
+			if ( $first_payment ) {
+				echo $first_payment->get_source_text(); //xss ok
 			} else {
 				printf(
 					'%s<br />%s', //xss ok
@@ -190,14 +180,14 @@ $payment = $subscription->get_first_payment();
 		</td>
 	</tr>
 
-	<?php if ( 's2member' === $payment->get_source() ) : ?>
+	<?php if ( 's2member' === $subscription->get_source() ) : ?>
 
 		<tr>
 			<th scope="row">
 				<?php esc_html_e( 'Period', 'pronamic_ideal' ); ?>
 			</th>
 			<td>
-				<?php echo esc_html( get_post_meta( $payment->get_id(), '_pronamic_payment_s2member_period', true ) ); ?>
+				<?php echo esc_html( get_post_meta( $subscription->get_id(), '_pronamic_subscription_s2member_period', true ) ); ?>
 			</td>
 		</tr>
 		<tr>
@@ -205,49 +195,7 @@ $payment = $subscription->get_first_payment();
 				<?php esc_html_e( 'Level', 'pronamic_ideal' ); ?>
 			</th>
 			<td>
-				<?php echo esc_html( get_post_meta( $payment->get_id(), '_pronamic_payment_s2member_level', true ) ); ?>
-			</td>
-		</tr>
-
-	<?php endif; ?>
-
-	<?php if ( 'wp-e-commerce' === $payment->get_source() ) : ?>
-
-		<tr>
-			<th scope="row">
-				<?php esc_html_e( 'Purchase ID', 'pronamic_ideal' ); ?>
-			</th>
-			<td>
-				<?php echo esc_html( get_post_meta( $payment->get_id(), '_pronamic_payment_wpsc_purchase_id', true ) ); ?>
-			</td>
-		</tr>
-		<tr>
-			<th scope="row">
-				<?php esc_html_e( 'Session ID', 'pronamic_ideal' ); ?>
-			</th>
-			<td>
-				<?php echo esc_html( get_post_meta( $payment->get_id(), '_pronamic_payment_wpsc_session_id', true ) ); ?>
-			</td>
-		</tr>
-
-	<?php endif; ?>
-
-	<?php if ( 'membership' === $payment->get_source() ) : ?>
-
-		<tr>
-			<th scope="row">
-				<?php esc_html_e( 'User ID', 'pronamic_ideal' ); ?>
-			</th>
-			<td>
-				<?php echo esc_html( get_post_meta( $payment->get_id(), '_pronamic_payment_membership_user_id', true ) ); ?>
-			</td>
-		</tr>
-		<tr>
-			<th scope="row">
-				<?php esc_html_e( 'Subscription ID', 'pronamic_ideal' ); ?>
-			</th>
-			<td>
-				<?php echo esc_html( get_post_meta( $payment->get_id(), '_pronamic_payment_membership_subscription_id', true ) ); ?>
+				<?php echo esc_html( get_post_meta( $subscription->get_id(), '_pronamic_subscription_s2member_level', true ) ); ?>
 			</td>
 		</tr>